Character Animations.
Created an animation controller and used the existing Third Person Character controller (starter pack) animations for idle and walk animations inside of the controller. Connected the transitions and created an "AnimState" parameter with int values of 0 (idle) and 1 (walking).

Inside the Character AI script, create a reference/instance of the animator and at the point in the script where the character translates towards the waypoint, I set the "AnimState" parameter int value to 1. As the character begins to move towards the waypoint, the walking animation kicks in.

Inside the "else" statement (when character isn't moving), set "AnimState" parameter back to 0 for idle.