Created another child under the Managers Game Object prefab
Created a "MissionManager" empty object that will hold a MissionManager.cs to manage status of the different missions that each farmer will perform at each level.

MissionManager will generate booleans to communicate with the Farmer GO speech bubble to know when to activate the speech bubble and eventually, which level dialogue to use.