Commit updates to DialogueManager_v2
Reverted back to Arrays for the dialogue containers. This time, using 2D arrays of 1 column (for now) and several rows. The rows will contain the dialogue that will be displayed in sequence.

Later, can use higher dimensional columns to represent the level that the string elements in that column belong to.